[native] Isolate form frame analysis
Move worklet scheduling, OpenCV detection, smoothing, and readiness into
a single hook with a small camera-facing interface.

Keep the form focused on capture and vote-entry orchestration.
